+++ tests/testthat/test-esp-make-provider.R
@@ -1,13 +1,8 @@
 test_that("Custom WMTS provider", {
-  skip_if_not_installed("terra")
-  skip_if_not_installed("png")
-  skip_on_os("mac")
-
   # Skip test as tiles sometimes are not available
   skip_on_cran()
   skip_if_offline()
 
-  segovia <- esp_get_prov("segovia", epsg = 3857)
   custom_wmts <- esp_make_provider(
     id = "wmts_test",
     q = "https://www.ign.es/wmts/ign-base?",
@@ -15,19 +10,21 @@ test_that("Custom WMTS provider", {
     layer = "IGNBaseTodo-nofondo"
   )
 
-  tile <- esp_get_tiles(segovia, type = custom_wmts)
-  expect_s4_class(tile, "SpatRaster")
+  # Ignore tilematrixset
+  custom_wmts2 <- esp_make_provider(
+    id = "wmts_test",
+    q = "https://www.ign.es/wmts/ign-base", # No ? at the end
+    service = "WMTS",
+    layer = "IGNBaseTodo-nofondo",
+    tileMatrixSet = "ignored"
+  )
+
+  expect_identical(custom_wmts, custom_wmts2)
 })
 
 test_that("Custom WMS provider", {
-  skip_if_not_installed("terra")
-  skip_if_not_installed("png")
-  skip_on_os("mac")
-
-  # Skip test as tiles sometimes are not available
   skip_on_cran()
   skip_if_offline()
-  segovia <- esp_get_prov("segovia", epsg = 3857)
 
   custom_wms_11 <- esp_make_provider(
     id = "wms_1.1",
@@ -38,19 +35,21 @@ test_that("Custom WMS provider", {
     layers = "geolog_cyl_litologia"
   )
 
+  # In pieces:
+  custom_wms_11_pieces <- validate_provider(custom_wms_11)
+  expect_null(ensure_null(custom_wms_11_pieces$crs))
+  expect_identical(ensure_null(custom_wms_11_pieces$srs), "EPSG:25830")
+
   custom_wms_13 <- esp_make_provider(
     id = "wms_1.3",
     q = "https://idecyl.jcyl.es/geoserver/ge/wms?",
     service = "WMS",
     version = "1.3.0",
-    crs = "EPSG:25830",
+    srs = "EPSG:25830",
     layers = "geolog_cyl_litologia"
   )
 
-  # Both works
-
-  tilewms1 <- esp_get_tiles(segovia, custom_wms_11, cache_dir = tempdir())
-  expect_s4_class(tilewms1, "SpatRaster")
-  tilewms13 <- esp_get_tiles(segovia, custom_wms_13, cache_dir = tempdir())
-  expect_s4_class(tilewms13, "SpatRaster")
+  custom_wms_13_pieces <- validate_provider(custom_wms_13)
+  expect_null(ensure_null(custom_wms_13_pieces$srs))
+  expect_identical(ensure_null(custom_wms_13_pieces$crs), "EPSG:25830")
 })