+++ tests/testthat/test_local/test-wms-providers.R
@@ -30,7 +30,7 @@ test_that("Test WMS png", {
 
   for (n in all_n) {
     tile <- try(
-      esp_get_tiles2(galicia, type = n, cache_dir = cdir),
+      esp_get_tiles(galicia, type = n, cache_dir = cdir),
       silent = TRUE
     )
     if (!inherits(tile, "try-error")) {