+++ tests/testthat/test_local/test-wmts-providers.R
@@ -30,7 +30,7 @@ test_that("Test WMTS png", {
 
   for (n in all_n) {
     tile <- try(
-      esp_get_tiles2(rioja, type = n, cache_dir = cdir),
+      esp_get_tiles(rioja, type = n, cache_dir = cdir),
       silent = TRUE
     )
     if (!inherits(tile, "try-error")) {